test(library): three clippy-1.97 fixes — CI's clippy is newer than local
unnecessary_cast (get() already returns u64), identity_op (spell the
D(4,4)=1 expectation directly, derivation in the comment), and an allow
for the 8-coordinate segment helper (the cell's own state shape). These
landed with the 395 batch and fail CI's clippy 1.97 on all three OSes;
local 1.95 doesn't carry the lints.

+++ cell80/tests/library/agentic-runtime.rs
@@ -523,7 +523,7 @@ fn rising_edge_step_fires_only_on_the_transition_to_one() {
     for (input, want_edge) in expected {
         let (edge, cell) = step(input, prev);
         assert_eq!(edge, want_edge, "input={input} prev={prev}");
-        prev = cell.get("prev").unwrap() as u64;
+        prev = cell.get("prev").unwrap();
         assert_eq!(prev, input); // prev always tracks the last input after run()
     }
 }
